==> Workflow/0-1_zincdata_handling.ipynb <==
import os

os.getcwd()

# ascii code 65 = "A"
# ord() / chr()
a = ord("A")
a

# total substance confirm
length = 0 

for x in range(66, 75):
    for y in range(65, 75):

        x_char = chr(x)
        y_char = chr(y)
        files = os.listdir(f"/home/dbsejrgus226/Huntington-s-Disease-Drug-Discovery/zinc_dataset/{x_char}{y_char}")

        
        for f in files:
            data = open(f"./zinc_dataset/{x_char}{y_char}/{f}", "r", encoding="UTF8")
            lines = data.readlines()
            ln = len(lines)
            length += ln
            data.close()


print(length)

# 1-1)each file is merged into one file
for x in range(66, 75):
    for y in range(65 ,75):
        
        x_char =chr(x)
        y_char =chr(y)
        outfilename = f"zinc_dataset{x_char}{y_char}.txt"
        files = os.listdir(f"/home/dbsejrgus226/Huntington-s-Disease-Drug-Discovery/zinc_dataset/{x_char}{y_char}")

        output = open(outfilename, "w")
        for f in files:
            result = open(f"./zinc_dataset/{x_char}{y_char}/{f}")
            for line in result:
                output.write(line)

            result.close()

        output.close()
        

# 1-2) check the total number of substances
length = 0 

x_char = chr(x)
y_char = chr(y)
files = os.listdir(f"/home/dbsejrgus226/Huntington-s-Disease-Drug-Discovery/zinc_dataset_2")

        
for f in files:
    data = open(f"./zinc_dataset_2/{f}", "r", encoding="UTF8")
    lines = data.readlines()
    ln = len(lines)
    length += ln
    data.close()


print(length)

# 2-1)each file is merged into one file
x_char = chr(x)
y_char = chr(y)
outfilename = "zinc_dataset_final.txt"
files = os.listdir("/home/dbsejrgus226/Huntington-s-Disease-Drug-Discovery/zinc_dataset_2")

output = open(outfilename, "w")
for f in files:
    result = open(f"./zinc_dataset_2/{f}")
    for line in result:
        output.write(line)

    result.close()

output.close()


# 2-2) check the total number of substances

data = open("zinc_dataset_final.txt", "r", encoding="UTF8")
lines = data.readlines()
ln = len(lines)
print(ln)
data.close()



import pandas as pd

data = pd.read_csv("zinc_dataset_final.txt", sep="\t", encoding="UTF-8")
data

selection = ["zinc_id", "smiles"]
data = data[selection]
data

data.isnull().sum()
