==> Workflow/0-2_choose_one_reported_Compound_Docking_to_Protein_ligand_complex_tweice_check.ipynb <==
import warnings
import subprocess
from pathlib import Path

import nglview as nv
from openbabel import pybel
from opencadd.structure.core import Structure

import pandas as pd


warnings.filterwarnings("ignore")
ob_log_handler = pybel.ob.OBMessageHandler()
pybel.ob.obErrorLog.SetOutputLevel(0)

com_data = pd.read_csv("./data/fak_pdbbind_dataset.csv", skiprows=1)

com_data

import requests
from opencadd.structure.core import Structure

def fetch_pdb_data(pdb_id):
    url = f"https://files.rcsb.org/download/{pdb_id}.pdb"
    response = requests.get(url)
    response.raise_for_status()
    with open(f"{pdb_id}.pdb", 'w') as file:
        file.write(response.text)
    return f"{pdb_id}.pdb"

structure_list = []
for i in com_data["PDB code"]:
    pdb_file = fetch_pdb_data(i)
    structure = Structure(pdb_file)
    structure_list.append(structure)

print(structure_list)


#for i in range(len(structure_list)):
#    protein = structure_list[i].select_atoms("protein")
#    protein.write(f"protein_{i}")

#def pdb_to_pdbqt(pdb_path, pdbqt_path, pH=7.4):
    """
    Convert a PDB file to a PDBQT file needed by docking programs of the AutoDock family.

    Parameters
    ----------
    pdb_path: str or pathlib.Path
        Path to input PDB file.
    pdbqt_path: str or pathlib.path
        Path to output PDBQT file.
    pH: float
        Protonation at given pH.
    """
    molecule = list(pybel.readfile("pdb", str(pdb_path)))[0]
    # add hydrogens at given pH
    molecule.OBMol.CorrectForPH(pH)
    molecule.addh()
    # add partial charges to each atom
    for atom in molecule.atoms:
        atom.OBAtom.GetPartialCharge()
    molecule.write("pdbqt", str(pdbqt_path), overwrite=True)
    return

# Convert protein to PDBQT format
for i in range(len(structure_list)):
    pdb_to_pdbqt(f"protein_{i}.pdb", f"protein_{i}.pdbqt")
    

# Zn27 smiles
smiles = "COc1ccc([C@@H]2CCCN2C(=O)Nc2cc(C(F)(F)F)ccc2N2CCOCC2)cc1"

def smiles_to_pdbqt(smiles, pdbqt_path, pH=7.4):
    """
    Convert a SMILES string to a PDBQT file needed by docking programs of the AutoDock family.

    Parameters
    ----------
    smiles: str
        SMILES string.
    pdbqt_path: str or pathlib.path
        Path to output PDBQT file.
    pH: float
        Protonation at given pH.
    """
    molecule = pybel.readstring("smi", smiles)
    # add hydrogens at given pH
    molecule.OBMol.CorrectForPH(pH)
    molecule.addh()
    # generate 3D coordinates
    molecule.make3D(forcefield="mmff94s", steps=10000)
    # add partial charges to each atom
    for atom in molecule.atoms:
        atom.OBAtom.GetPartialCharge()
    molecule.write("pdbqt", str(pdbqt_path), overwrite=True)
    return

smiles_to_pdbqt(smiles, "ligand_Zn27.pdbqt")

pocketsize_list = []
pocketcenter_list = []
for i in range(len(com_data)):
    ligand_resname = com_data["Ligand Name"][i]
    ligand = structure_list[i].select_atoms(f"resname {ligand_resname}")
    pocket_center = (ligand.positions.max(axis=0) + ligand.positions.min(axis=0)) / 2
    pocket_size = ligand.positions.max(axis=0) - ligand.positions.min(axis=0) +5
    pocketsize_list.append(pocket_size)
    pocketcenter_list.append(pocket_center)

pocketcenter_list

pocketsize_list

pocketcenter_list = pd.Series(pocketsize_list)
pocketsize_list = pd.Series(pocketsize_list)

pocketinfo = pd.concat([pocketcenter_list, pocketsize_list], axis=1)
pocketinfo

ligandname = []
for i in com_data["Ligand Name"]:
    ligandname.append(i)

ligandname = pd.Series(ligandname)
ligandname

pocketinfo = pd.concat([ligandname, pocketinfo], axis=1)
pocketinfo

pocketinfo.columns = ["Ligand Name", "pocket center", "pocket size"]
pocketinfo

pocketinfo["pocket center"][0][0]

def run_smina(
    ligand_path, protein_path, out_path, pocket_center, pocket_size, num_poses=10, exhaustiveness=8
):
    """
    Perform docking with Smina.

    Parameters
    ----------
    ligand_path: str or pathlib.Path
        Path to ligand PDBQT file that should be docked.
    protein_path: str or pathlib.Path
        Path to protein PDBQT file that should be docked to.
    out_path: str or pathlib.Path
        Path to which docking poses should be saved, SDF or PDB format.
    pocket_center: iterable of float or int
        Coordinates defining the center of the binding site.
    pocket_size: iterable of float or int
        Lengths of edges defining the binding site.
    num_poses: int
        Maximum number of poses to generate.
    exhaustiveness: int
        Accuracy of docking calculations.

    Returns
    -------
    output_text: str
        The output of the Smina calculation.
    """
    output_text = subprocess.check_output(
        [
            # In Max   #########
            #"./smina.osx.12",

            # In Linux #########
            "./smina.linux",
            
            "--ligand",
            str(ligand_path),
            "--receptor",
            str(protein_path),
            "--out",
            str(out_path),
            "--center_x",
            str(pocket_center[0]),
            "--center_y",
            str(pocket_center[1]),
            "--center_z",
            str(pocket_center[2]),
            "--size_x",
            str(pocket_size[0]),
            "--size_y",
            str(pocket_size[1]),
            "--size_z",
            str(pocket_size[2]),
            "--num_modes",
            str(num_poses),
            "--exhaustiveness",
            str(exhaustiveness),
        ],
        universal_newlines=True,  # needed to capture output text
    )
    return output_text

Zn27docking = open("Zn27docking_score_twice.txt", mode="w")

for i in range(7):
    results = run_smina(
        "./data/pdbqt_file/ligand_Zn27.pdbqt",
        f"./data/pdbqt_file/protein_{i}.pdbqt",
        f"docking_poses_{i}.sdf",
        pocketinfo['pocket center'][i],
        pocketinfo['pocket size'][i]
    )
    Zn27docking.write(results)
    print(results)

Zn27docking.close()

# 5번쨰 protein good

# 6번쨰 단백질 결과가 이상해서 다시 해본 것이며 동일하게ㅐ 이상하게 나옴
