==> Workflow/1-2_Lipinski.data_merging.ipynb <==
import numpy as np
from rdkit import Chem
from rdkit.Chem import Descriptors, Lipinski

# 1_lipinski.py를 통해 아래 데이터 구축 완료

import pandas as pd
data_Mt = pd.read_csv("Molecular_Weight.csv", encoding="UTF-8", usecols=["smiles", "zinc_id", "MolWt"])
data_LogP = pd.read_csv("LogP.csv", encoding="UTF-8", usecols=["LogP"])
data_HD = pd.read_csv("HDonors.csv", encoding="UTF-8", usecols=["HDonors"])
data_HA = pd.read_csv("HAcceptors.csv", encoding="UTF-8", usecols=["HAcceptors"])

lipinski_data = pd.concat([data_Mt, data_LogP, data_HD, data_HA], axis=1)
lipinski_data

invalid_MolWt = lipinski_data[lipinski_data["MolWt"] > 500].index
lipinski_data.drop(invalid_MolWt, inplace=True)
lipinski_data

lipinski_data = lipinski_data.reset_index(drop=True)
lipinski_data

invalid_MolWt = lipinski_data[lipinski_data["LogP"] > 5].index
lipinski_data.drop(invalid_MolWt, inplace=True)
lipinski_data

lipinski_data = lipinski_data.reset_index(drop=True)
lipinski_data

invalid_MolWt = lipinski_data[lipinski_data["HDonors"] > 5].index
lipinski_data.drop(invalid_MolWt, inplace=True)
lipinski_data

lipinski_data = lipinski_data.reset_index(drop=True)
lipinski_data

invalid_MolWt = lipinski_data[lipinski_data["HAcceptors"] > 10].index
lipinski_data.drop(invalid_MolWt, inplace=True)
lipinski_data

lipinski_data = lipinski_data.reset_index(drop=True)
lipinski_data

rowdata = pd.concat([data_Mt, data_LogP, data_HD, data_HA], axis=1)
a = print(len(rowdata))
b = print(len(lipinski_data))
c = len(rowdata) - len(lipinski_data)
print(c)

# lipinski filtering 전 10725143개
# filtering 후 10703772개
# 21371개가 lipinski five rule에 의해 제거됨

lipinski_data.to_csv("lipinski_data.csv")
