==> Workflow/4-2_check_the_zn27_cluster_feature.ipynb <==
import pandas as pd

total_data = pd.read_csv("clustered_data_by_kmeans.csv", index_col=0, low_memory=False)
total_data.head()

len(total_data)

total_data.tail()

total_data.drop(10301988, axis=0, inplace=True)
total_data.tail()

len(total_data)

final_data = pd.read_csv("final_data_added_lipinski.csv", low_memory=False)
final_data.head()

len(final_data)

final_data["cluster"] = total_data["cluster"]
final_data.head()

average_mw = final_data["MolWt"].mean()
average_logp = final_data["LogP"].mean()
average_hd = final_data["HDonors"].mean()
average_ha = final_data["HAcceptors"].mean()

print(average_mw)
print(average_logp)
print(average_hd)
print(average_ha)

cluster_6_data = final_data[final_data["cluster"] == 6].reset_index(drop=True)

cluster_6_data

cluster_6_average_mw = cluster_6_data["MolWt"].mean()
cluster_6_average_logp = cluster_6_data["LogP"].mean()
cluster_6_average_hd = cluster_6_data["HDonors"].mean()
cluster_6_average_ha = cluster_6_data["HAcceptors"].mean()

print(cluster_6_average_mw)
print(cluster_6_average_logp)
print(cluster_6_average_hd)
print(cluster_6_average_ha)

cluster 6에 속한 molecules은 전체 평균보다 MW 및 LogP가 각각 대략 100과 1 만큼 높으며 이것은 둘다 평균보다 1.3배 차이를 보입니다. HDonor와 HAcceptor는 평균과 거의 비슷합니다.

import time
import random
from pathlib import Path

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from rdkit import Chem
from rdkit import DataStructs
from rdkit.ML.Cluster import Butina 
from rdkit.Chem import Draw
from rdkit.Chem import rdFingerprintGenerator
from rdkit.Chem import PandasTools

from tqdm import tqdm

import numpy as np
from rdkit import Chem
from rdkit.Chem import Descriptors, Lipinski

zinc27 = Chem.MolFromSmiles("COc1ccc([C@@H]2CCCN2C(=O)Nc2cc(C(F)(F)F)ccc2N2CCOCC2)cc1")
zinc27

desc_MolWt = Descriptors.MolWt(zinc27)
desc_MolLogP = Descriptors.MolLogP(zinc27)
desc_NumHDonors = Descriptors.NumHDonors(zinc27)
desc_NumHAcceptors = Descriptors.NumHAcceptors(zinc27)

row = np.array([desc_MolWt, desc_MolLogP, desc_NumHDonors, desc_NumHAcceptors])
row

row = pd.DataFrame(row.reshape(1, -1), columns=["MolWt", "LogP", "HDonors", "HAcceptors"])
row
