==> Workflow/5_check_common_data_between_lipinski_and_tanimoto.ipynb <==
import pandas as pd

# tanimoto similarity에 의해 선정된 화합물들과 lipinski 4가지 특징을 이용하여 K-means clustering을 수행하여 선정된 화합물 사이 공통 화합물 확인을 위한 과정정

lipinski_data = pd.read_csv("final_data_with_cluster.csv")
lipinski_data.head()

tanimoto_data = pd.read_csv("similarity_with_zn27_by_tanimoto_maccs.csv")
tanimoto_data.head()

lipinski_data.drop(columns=["Unnamed: 0"], inplace=True)
lipinski_data.head()
print(len(lipinski_data))

lipinski_data.head()

print(len(lipinski_data))

tanimoto_data.drop(columns=["Unnamed: 0"], inplace=True)

tanimoto_data.head()

print(len(tanimoto_data))

common_data = []
for id in lipinski_data["zinc_id"]:
    if id in tanimoto_data["zinc_id"].values:
        common_data.append(id)

print(common_data)
print(len(common_data))

commom_lipinski_data = lipinski_data[lipinski_data["zinc_id"].isin(common_data)].reset_index(drop=True)
commom_lipinski_data

common_tanimoto_data = tanimoto_data[tanimoto_data["zinc_id"].isin(common_data)].reset_index(drop=True)
common_tanimoto_data

common_tanimoto_data.drop(columns=["zinc_id", "smiles"], inplace=True)

final_common_data_between_lipinski_and_tanimoto = pd.concat([commom_lipinski_data, common_tanimoto_data], axis=1)
final_common_data_between_lipinski_and_tanimoto

final_common_data_between_lipinski_and_tanimoto.to_csv("final_common_data_between_lipinski_and_tanimoto.csv")
