==> Workflow/9-3_calculate_SAScore.ipynb <==
from rdkit import Chem
from rdkit.Chem.Draw import IPythonConsole
from rdkit.Chem import rdDepictor
rdDepictor.SetPreferCoordGen(True)
import rdkit
print(rdkit.__version__)

import sys
import os
sys.path.append(os.path.join(os.environ["CONDA_PREFIX"],"share","RDKit","Contrib"))

from SA_Score import sascorer
from NP_Score import npscorer

import pandas as pd
data = pd.read_csv("./merge_data_with_lipinski_tanimoto_and_common.csv", usecols=["smiles"])
data.head()

len(data)

sascores = []
for i in range(len(data)):
    mol = Chem.MolFromSmiles(data["smiles"][i])
    sascores.append(sascorer.calculateScore(mol))

data["SAScore"] = sascores

ZN27 = "COc1ccc([C@@H]2CCCN2C(=O)Nc2cc(C(F)(F)F)ccc2N2CCOCC2)cc1"
mol = Chem.MolFromSmiles(ZN27)
SAS = sascorer.calculateScore(mol)
SAS

data.head()

ZN27_SAS = pd.DataFrame({"smiles":[ZN27], "SAScore":[SAS]})
ZN27_SAS.head()

data.to_csv("./SAScore_results.csv")

ZN27_SAS.to_csv("./ZN27_SAScore.csv")
