==> Workflow/12_check_the_top10_groups.ipynb <==
import pandas as pd

rank10 = pd.read_csv("./rank_top_10_molecules.csv", index_col=0)
rank10.drop(columns=["Unnamed: 0"], inplace=True)
rank10.head()

groups = pd.read_csv("./merge_data_with_lipinski_tanimoto_and_common(only_smiles).csv", index_col=0)
groups.head()

len(groups)

only_common = pd.read_csv("./final_common_data_between_lipinski_and_tanimoto.csv", index_col=0)
only_common.head()

lst_1 = []

for i in range(len(rank10)):
    for x in range(len(only_common)):
        if rank10["smiles"].iloc[i] == only_common["smiles"].iloc[x]:
            lst_1.append(x)

lst_1

lst_2 = []

for i in range(len(rank10)):
    for y in range(len(groups)):
        if rank10["smiles"].iloc[i] == groups["smiles"].iloc[y]:
            lst_2.append(y)

lst_2

# 0 <= x <= 65125 (index 기준) : K-means clustering group
# 65126 <= x <= 70245 : Tanimoto similarity group

k_means_lst = []
tanimoto_lst = []

for i in lst_2:
    if 0 <= i <= 65125:
        k_means_lst.append(i)
    else:
        tanimoto_lst.append(i)

print(len(k_means_lst))
print(len(tanimoto_lst))

### Comparing top 10 with ZN27

import pandas as pd

top10 = pd.read_csv("./rank_top_10_molecules.csv", index_col=0)
top10.head()
